# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stats {
  memories: number;
  cafes: number;
  books: number;
  movies: number;
  kitchen: number;
  reviews: number;
  wishes: number;
}

export interface TimelineItem {
  id: string;
  section: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  createdAt: string;
}

export interface Memory {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  mediaType?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  isFavorite: boolean | null;
  createdAt: string;
}

export interface MemoryInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  mediaUrl?: string;
  mediaType?: string;
  date?: string;
  /** @nullable */
  isFavorite?: boolean | null;
}

export interface MemoryUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  mediaUrl?: string;
  mediaType?: string;
  date?: string;
  /** @nullable */
  isFavorite?: boolean | null;
}

export interface Cafe {
  id: string;
  name: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  whatWeAte?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  reflection?: string | null;
  /** @nullable */
  wouldVisitAgain: boolean | null;
  createdAt: string;
}

export interface CafeInput {
  /** @minLength 1 */
  name: string;
  location?: string;
  date?: string;
  rating?: number;
  whatWeAte?: string;
  photoUrl?: string;
  reflection?: string;
  /** @nullable */
  wouldVisitAgain?: boolean | null;
}

export interface CafeUpdate {
  /** @minLength 1 */
  name?: string;
  location?: string;
  date?: string;
  rating?: number;
  whatWeAte?: string;
  photoUrl?: string;
  reflection?: string;
  /** @nullable */
  wouldVisitAgain?: boolean | null;
}

export interface Book {
  id: string;
  title: string;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  quote?: string | null;
  /** @nullable */
  reflection?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  isStayed?: boolean | null;
  createdAt: string;
}

export interface BookInput {
  /** @minLength 1 */
  title: string;
  author?: string;
  status?: string;
  rating?: number;
  quote?: string;
  reflection?: string;
  coverUrl?: string;
  /** @nullable */
  isStayed?: boolean | null;
}

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  author?: string;
  status?: string;
  rating?: number;
  quote?: string;
  reflection?: string;
  coverUrl?: string;
  /** @nullable */
  isStayed?: boolean | null;
}

export interface Movie {
  id: string;
  title: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  whatItReminded?: string | null;
  /** @nullable */
  posterUrl?: string | null;
  createdAt: string;
}

export interface MovieInput {
  /** @minLength 1 */
  title: string;
  type?: string;
  genre?: string;
  rating?: number;
  whatItReminded?: string;
  posterUrl?: string;
}

export interface MovieUpdate {
  /** @minLength 1 */
  title?: string;
  type?: string;
  genre?: string;
  rating?: number;
  whatItReminded?: string;
  posterUrl?: string;
}

export interface Wish {
  id: string;
  /** @nullable */
  emoji?: string | null;
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  isDone: boolean | null;
  createdAt: string;
}

export interface WishInput {
  emoji?: string;
  /** @minLength 1 */
  title: string;
  notes?: string;
  /** @nullable */
  isDone?: boolean | null;
}

export interface WishUpdate {
  emoji?: string;
  /** @minLength 1 */
  title?: string;
  notes?: string;
  /** @nullable */
  isDone?: boolean | null;
}

export interface Capsule {
  id: string;
  title: string;
  message: string;
  unlockAt: string;
  /** @nullable */
  isUnlocked: boolean | null;
  createdAt: string;
}

export interface CapsuleInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  message: string;
  unlockAt: string;
}

export interface KitchenEntry {
  id: string;
  title: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  recipe?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  wouldMakeAgain?: boolean | null;
  /** @nullable */
  date?: string | null;
  createdAt: string;
}

export interface KitchenEntryInput {
  /** @minLength 1 */
  title: string;
  type?: string;
  recipe?: string;
  photoUrl?: string;
  rating?: number;
  mood?: string;
  notes?: string;
  /** @nullable */
  wouldMakeAgain?: boolean | null;
  date?: string;
}

export interface KitchenEntryUpdate {
  /** @minLength 1 */
  title?: string;
  type?: string;
  recipe?: string;
  photoUrl?: string;
  rating?: number;
  mood?: string;
  notes?: string;
  /** @nullable */
  wouldMakeAgain?: boolean | null;
  date?: string;
}

export interface Review {
  id: string;
  subject: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  vibeCheck?: string | null;
  review: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  wouldRecommend?: boolean | null;
  /** @nullable */
  date?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  /** @minLength 1 */
  subject: string;
  category?: string;
  rating?: number;
  vibeCheck?: string;
  /** @minLength 1 */
  review: string;
  photoUrl?: string;
  /** @nullable */
  wouldRecommend?: boolean | null;
  date?: string;
}

export interface ReviewUpdate {
  /** @minLength 1 */
  subject?: string;
  category?: string;
  rating?: number;
  vibeCheck?: string;
  /** @minLength 1 */
  review?: string;
  photoUrl?: string;
  /** @nullable */
  wouldRecommend?: boolean | null;
  date?: string;
}

export type ListMemoriesParams = {
favorite?: string;
search?: string;
};

export type ListBooksParams = {
status?: string;
};

export type ListKitchenEntriesParams = {
type?: string;
};