# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BloomBook API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cafes:
    get:
      operationId: listCafes
      tags: [cafes]
      responses:
        "200":
          description: Cafes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cafe"
    post:
      operationId: createCafe
      tags: [cafes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CafeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cafe"
  /cafes/{id}:
    get:
      operationId: getCafe
      tags: [cafes]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Cafe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cafe"
    patch:
      operationId: updateCafe
      tags: [cafes]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CafeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cafe"
    delete:
      operationId: deleteCafe
      tags: [cafes]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted

  # ── BOOKS ──────────────────────────────────────────────────────────────
  /books:
    get:
      operationId: listBooks
      tags: [books]
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    patch:
      operationId: updateBook
      tags: [books]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted

  # ── MOVIES ─────────────────────────────────────────────────────────────
...
  /capsules:
    get:
      operationId: listCapsules
      tags: [capsules]
      responses:
        "200":
          description: Capsules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Capsule"
    post:
      operationId: createCapsule
      tags: [capsules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapsuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capsule"
  /capsules/{id}:
    get:
      operationId: getCapsule
      tags: [capsules]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Capsule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capsule"
    delete:
      operationId: deleteCapsule
      tags: [capsules]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
  /capsules/{id}/unlock:
    patch:
      operationId: unlockCapsule
      tags: [capsules]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capsule"

  # ── KITCHEN ────────────────────────────────────────────────────────────
  /kitchen:
    get:
      operationId: listKitchenEntries
      tags: [kitchen]
      parameters:
        - in: query
          name: type
          schema:
            type: string
      responses:
        "200":
          description: Kitchen entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KitchenEntry"
    post:
      operationId: createKitchenEntry
      tags: [kitchen]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KitchenEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenEntry"
  /kitchen/{id}:
    get:
      operationId: getKitchenEntry
      tags: [kitchen]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Kitchen entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenEntry"
    patch:
      operationId: updateKitchenEntry
      tags: [kitchen]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KitchenEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenEntry"
    delete:
      operationId: deleteKitchenEntry
      tags: [kitchen]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted

  # ── REVIEWS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Cafe:
      type: object
      required: [id, name, wouldVisitAgain, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        whatWeAte:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        reflection:
          type: ["string", "null"]
        wouldVisitAgain:
          type: ["boolean", "null"]
        createdAt:
          type: string
    CafeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
        date:
          type: string
        rating:
          type: integer
        whatWeAte:
          type: string
        photoUrl:
          type: string
        reflection:
          type: string
        wouldVisitAgain:
          type: ["boolean", "null"]
    CafeUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
        date:
          type: string
        rating:
          type: integer
        whatWeAte:
          type: string
        photoUrl:
          type: string
        reflection:
          type: string
        wouldVisitAgain:
          type: ["boolean", "null"]

    # ── Book ──────────────────────────────────────────────────────────────
    Book:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        quote:
          type: ["string", "null"]
        reflection:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isStayed:
          type: ["boolean", "null"]
        createdAt:
          type: string
    BookInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        author:
          type: string
        status:
          type: string
        rating:
          type: integer
        quote:
          type: string
        reflection:
          type: string
        coverUrl:
          type: string
        isStayed:
          type: ["boolean", "null"]
    BookUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        author:
          type: string
        status:
          type: string
        rating:
          type: integer
        quote:
          type: string
        reflection:
          type: string
        coverUrl:
          type: string
        isStayed:
          type: ["boolean", "null"]

    # ── Movie ─────────────────────────────────────────────────────────────
...
    Capsule:
      type: object
      required: [id, title, message, unlockAt, isUnlocked, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        unlockAt:
          type: string
        isUnlocked:
          type: ["boolean", "null"]
        createdAt:
          type: string
    CapsuleInput:
      type: object
      required: [title, message, unlockAt]
      properties:
        title:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        unlockAt:
          type: string

    # ── KitchenEntry ──────────────────────────────────────────────────────
    KitchenEntry:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: ["string", "null"]
        recipe:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        mood:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        wouldMakeAgain:
          type: ["boolean", "null"]
        date:
          type: ["string", "null"]
        createdAt:
          type: string
    KitchenEntryInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
        recipe:
          type: string
        photoUrl:
          type: string
        rating:
          type: integer
        mood:
          type: string
        notes:
          type: string
        wouldMakeAgain:
          type: ["boolean", "null"]
        date:
          type: string
    KitchenEntryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
        recipe:
          type: string
        photoUrl:
          type: string
        rating:
          type: integer
        mood:
          type: string
        notes:
          type: string
        wouldMakeAgain:
          type: ["boolean", "null"]
        date:
          type: string

    # ── Review ────────────────────────────────────────────────────────────
...