# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  /** @minLength 1 */
  title: string;
  author?: string;
  status?: string;
  rating?: number;
  quote?: string;
  reflection?: string;
  coverUrl?: string;
  /** @nullable */
  isStayed?: boolean | null;
}