# lib/api-zod/src/generated/types/cafe.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Cafe {
  id: string;
  name: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  whatWeAte?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  reflection?: string | null;
  /** @nullable */
  wouldVisitAgain: boolean | null;
  createdAt: string;
}