# lib/api-zod/src/generated/types/cafeUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CafeUpdate {
  /** @minLength 1 */
  name?: string;
  location?: string;
  date?: string;
  rating?: number;
  whatWeAte?: string;
  photoUrl?: string;
  reflection?: string;
  /** @nullable */
  wouldVisitAgain?: boolean | null;
}