# lib/api-zod/src/generated/types/capsule.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Capsule {
  id: string;
  title: string;
  message: string;
  unlockAt: string;
  /** @nullable */
  isUnlocked: boolean | null;
  createdAt: string;
}