# lib/api-zod/src/generated/types/kitchenEntryInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KitchenEntryInput {
  /** @minLength 1 */
  title: string;
  type?: string;
  recipe?: string;
  photoUrl?: string;
  rating?: number;
  mood?: string;
  notes?: string;
  /** @nullable */
  wouldMakeAgain?: boolean | null;
  date?: string;
}